refactor(cli): use multiprocessing module to spawn worker processes
Closes #97
